requirements.txt
numpy
pandas
python-binance
ta-lib
websocket-client
